+++ .ncurc.js
@@ -4,12 +4,21 @@ module.exports = {
     '@types/react-native',
     'message2call',
     'react',
+    'react-native',
+    'react-native-pager-view',
+    'react-native-navigation',
   ],
+
+  // target: 'newest',
+  // filter: [
+  //   'react-native-navigation',
+  // ],
 
   // target: 'patch',
   // filter: [
   //   '@types/react-native',
   //   'react',
+  //   'react-native',
   //   '@react-native/metro-config',
   // ],
 }