+++ src/screens/Home/Views/Setting/settings/Basic/IsAlwaysKeepStatusbarHeight.tsx
@@ -10,14 +10,18 @@ import CheckBoxItem from '../../components/CheckBoxItem'
 
 export default memo(() => {
   const t = useI18n()
-  const isPlayHighQuality = useSettingValue('player.isPlayHighQuality')
-  const setPlayHighQuality = (isPlayHighQuality: boolean) => {
-    updateSetting({ 'player.isPlayHighQuality': isPlayHighQuality })
+  const val = useSettingValue('common.alwaysKeepStatusbarHeight')
+  const update = (alwaysKeepStatusbarHeight: boolean) => {
+    updateSetting({ 'common.alwaysKeepStatusbarHeight': alwaysKeepStatusbarHeight })
   }
 
   return (
     <View style={styles.content}>
-      <CheckBoxItem check={isPlayHighQuality} onChange={setPlayHighQuality} label={t('setting_play_quality')} />
+      <CheckBoxItem
+        check={val}
+        label={t('setting_basic_always_keep_statusbar_height')}
+        helpDesc={t('setting_basic_always_keep_statusbar_height_tip')}
+        onChange={update} />
     </View>
   )
 })
@@ -26,6 +30,6 @@ export default memo(() => {
 const styles = createStyle({
   content: {
     marginTop: 5,
+    marginBottom: 15,
   },
 })
-