+++ src/screens/PlayDetail/Horizontal/Lyric.tsx
@@ -207,6 +207,7 @@ export default () => {
         }, 100)
       }
     })
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [lyricLines])
 
   useEffect(() => {
@@ -224,6 +225,7 @@ export default () => {
       delayScrollTimeout.current = null
       handleScrollToActive()
     }, 600)
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [line])
 
   useEffect(() => {