+++ src/types/app_setting.d.ts
@@ -67,6 +67,11 @@ declare global {
       'common.useSystemFileSelector': boolean
 
       /**
+       * 总是保留状态栏高度
+       */
+      'common.alwaysKeepStatusbarHeight': boolean
+
+      /**
        * 主题id
        */
       'theme.id': string
@@ -97,9 +102,14 @@ declare global {
       'theme.fontShadow': boolean
 
       /**
-           * 启动时自动播放歌曲
-           */
+       * 启动时自动播放歌曲
+       */
       'player.startupAutoPlay': boolean
+
+      /**
+       * 启动后打开歌曲详细界面
+       */
+      'player.startupPushPlayDetailScreen': boolean
 
       /**
        * 切歌模式
@@ -107,9 +117,9 @@ declare global {
       'player.togglePlayMethod': 'listLoop' | 'random' | 'list' | 'singleLoop' | 'none'
 
       /**
-       * 是否优先播放320k音质
+       * 优先播放的音质
        */
-      'player.isPlayHighQuality': boolean
+      'player.playQuality': LX.Quality
 
       /**
        * 启动软件时是否恢复上次播放进度