+++ CHANGELOG.md
@@ -6,6 +6,89 @@ Project versioning adheres to [Semantic Versioning](http://semver.org/).
 Commit convention is based on [Conventional Commits](http://conventionalcommits.org).
 Change log format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.5.0](https://github.com/lyswhut/lx-music-mobile/compare/v1.4.2...v1.5.0) - 2024-08-03
+
+我们发布了关于 LX Music 项目发展调整与新项目计划的说明，
+详情看： https://github.com/lyswhut/lx-music-desktop/issues/1912
+
+### 新增
+
+- 新增重复歌曲列表，可以方便移除我的列表中的重复歌曲，此列表会列出目标列表里歌曲名相同的歌曲，可在“我的列表”里的列表名菜单中使用（注：该功能与PC端的区别是可以点击歌曲名多选删除）
+- 新增打开当前歌曲详情页菜单，可以在歌曲菜单中使用
+
+### 修复
+
+- 修复潜在桌面歌词导致的崩溃问题
+
+### 其他
+
+- 更新 React native 到 v0.73.9
+- 更新 exoplayer 到 v1.4.0
+
+## [1.4.2](https://github.com/lyswhut/lx-music-mobile/compare/v1.4.1...v1.4.2) - 2024-06-01
+
+我们发布了关于 LX Music 项目发展调整与新项目计划的说明，
+详情看： https://github.com/lyswhut/lx-music-desktop/issues/1912
+
+### 修复
+
+- 修复数据存储管理在移除数据时可能出现移除失败的问题
+
+## [1.4.1](https://github.com/lyswhut/lx-music-mobile/compare/v1.4.0...v1.4.1) - 2024-06-01
+
+我们发布了关于 LX Music 项目发展调整与新项目计划的说明，
+详情看： https://github.com/lyswhut/lx-music-desktop/issues/1912
+
+### 修复
+
+- 修复播放详情页歌词滚动问题（#518）
+
+## [1.4.0](https://github.com/lyswhut/lx-music-mobile/compare/v1.3.0...v1.4.0) - 2024-06-01
+
+我们发布了关于 LX Music 项目发展调整与新项目计划的说明，
+详情看： https://github.com/lyswhut/lx-music-desktop/issues/1912
+
+### 新增
+
+- 新增 设置-基本设置-启动后打开播放详情界面 设置，默认关闭（#502 @mingcc7）
+
+### 修复
+
+- 修复重复的数据初始化调用
+- 修复导入歌单时可能会导致歌单数据存储异常的问题（#500）
+
+### 变更
+
+- 设置-播放设置-优先播放320k音质选项改为“优先播放的音质”，允许选择更高优先播放的音质，如果歌曲及音源支持的话（#487）
+
+### 其他
+
+- 更新 React native 到 v0.73.8
+
+## [1.3.0](https://github.com/lyswhut/lx-music-mobile/compare/v1.2.0...v1.3.0) - 2024-04-14
+
+### 新增
+
+- 新增棕色主题“泥牛入海”
+- 新增设置-基本设置-总是保留状态栏高度设置，如果在你的设备上出现软件可交互内容与状态栏内容显示重叠的情况，可以启用该设置以始终为系统状态栏保留空间
+- 新增在线自定义源导入功能，允许通过http/https链接导入自定义源
+
+### 优化
+
+- 不再丢弃kg源逐行歌词（@helloplhm-qwq）
+- 支持kw源排行榜显示大小（revert @Folltoshe #1460）
+- 优化本地歌曲换源匹配机制
+
+### 修复
+
+- 修复mg歌词在某些情况下获取失败的问题
+- 修复mg歌单搜索（@helloplhm-qwq）
+- 修复kg最新评论无法获取的问题（@helloplhm-qwq）
+
+### 其他
+
+- 更新 React native 到 v0.73.6
+
 ## [1.2.0](https://github.com/lyswhut/lx-music-mobile/compare/v1.1.1...v1.2.0) - 2024-02-01
 
 提前祝大家新年快乐！